Move TelemetryApiFixture to TestFramework proj
