+++ test/TestFramework/Microsoft.ServiceFabric.TestFramework.csproj
@@ -11,6 +11,7 @@
     <PackageReference Include="Microsoft.ServiceFabric" />
     <PackageReference Include="Moq" />
     <PackageReference Include="xunit.v3.assert" />
+    <PackageReference Include="xunit.v3" /> 
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Constants\Microsoft.ServiceFabric.Constants.csproj" />
